moved the weaviate vector db class inside the vector_db folder, and updated the test case
